scene_breakdown/requirements.txt
openai==2.6.1
python-dotenv==1.0.1
httpx>=0.27.2
tqdm>=4.66.4